refactor(budget): migrate budget system from quota to currency units
Implements ADR budget-currency-based: all budget/consumed/remain fields
now store company billing_currency amounts instead of quota integers.

Backend:
- Schema: BIGINT → NUMERIC(18,6) for budget columns
- Go types: int64 → float64 across domain/store/pkg layers
- Ingest: spend = Σ segments.DisplayAmount (currency)
- Chain: GatewayChainRemain returns uncapped bool, no MaxInt64 sentinel
- Uncapped remain → NULL (skip persist), not sentinel value
- SumAmountByDepartment uses SUM(display_amount)
- ConsumedDrift uses epsilon comparison for float64

Frontend:
- Remove displayToQuota/quotaToDisplay from budget paths
- Use formatMoney for budget/consumed display
- Submit budget amounts directly as currency values

Seed:
- Remove seedQuota multiplier, store currency values directly

+++ apps/backend/internal/domain/budget/service.go
@@ -13,10 +13,10 @@ import (
 
 type Service interface {
 	GetTree(ctx context.Context) ([]types.BudgetNode, error)
-	UpdateNode(ctx context.Context, id uuid.UUID, budget int64, reservedPool *int64) (types.BudgetNode, error)
+	UpdateNode(ctx context.Context, id uuid.UUID, budget float64, reservedPool *float64) (types.BudgetNode, error)
 	ListMemberBudgets(ctx context.Context, deptID uuid.UUID) ([]types.MemberBudget, error)
-	UpdateMemberBudget(ctx context.Context, memberID uuid.UUID, personalBudget int64) (types.MemberBudget, error)
-	ApplyAverageBudget(ctx context.Context, deptID uuid.UUID, personalBudget int64, recursive bool) error
+	UpdateMemberBudget(ctx context.Context, memberID uuid.UUID, personalBudget float64) (types.MemberBudget, error)
+	ApplyAverageBudget(ctx context.Context, deptID uuid.UUID, personalBudget float64, recursive bool) error
 	ListProjects(ctx context.Context) ([]types.Project, error)
 	CreateProject(ctx context.Context, project types.Project) (types.Project, error)
 	UpdateProject(ctx context.Context, id uuid.UUID, patch types.UpdateProjectInput) (types.Project, error)
@@ -27,7 +27,7 @@ type Service interface {
 	CreateAlert(ctx context.Context, rule types.AlertRule) (types.AlertRule, error)
 	UpdateAlert(ctx context.Context, id uuid.UUID, patch types.AlertRule) (types.AlertRule, error)
 	DeleteAlert(ctx context.Context, id uuid.UUID) error
-	GetProjectMemberConsumed(ctx context.Context, projectID uuid.UUID) (map[uuid.UUID]int64, error)
+	GetProjectMemberConsumed(ctx context.Context, projectID uuid.UUID) (map[uuid.UUID]float64, error)
 	MemberSummary(ctx context.Context, memberID uuid.UUID) (types.MemberBudgetSummary, error)
 }
