+++ apps/backend/internal/domain/budget/tree_query.go
@@ -43,7 +43,7 @@ func (s *service) ListMemberBudgets(ctx context.Context, deptID uuid.UUID) ([]ty
 	return memberBudgets, nil
 }
 
-func (s *service) GetProjectMemberConsumed(ctx context.Context, projectID uuid.UUID) (map[uuid.UUID]int64, error) {
+func (s *service) GetProjectMemberConsumed(ctx context.Context, projectID uuid.UUID) (map[uuid.UUID]float64, error) {
 	projects, err := s.store.Budget().Projects(ctx)
 	if err != nil {
 		return nil, err
@@ -59,7 +59,7 @@ func (s *service) GetProjectMemberConsumed(ctx context.Context, projectID uuid.U
 		return nil, domain.NotFound("Project not found")
 	}
 	if len(target.MemberIDs) == 0 {
-		return make(map[uuid.UUID]int64), nil
+		return make(map[uuid.UUID]float64), nil
 	}
 
 	deptID := target.OwnerDepartmentID
@@ -75,7 +75,7 @@ func (s *service) GetProjectMemberConsumed(ctx context.Context, projectID uuid.U
 	}
 	consumedRepo := s.store.BudgetConsumed()
 
-	result := make(map[uuid.UUID]int64, len(target.MemberIDs))
+	result := make(map[uuid.UUID]float64, len(target.MemberIDs))
 	for _, memberID := range target.MemberIDs {
 		sum, err := pkgbudget.SumProjectMemberKeyConsumedFromRepo(
 			ctx, consumedRepo, keys, projectID, memberID, periodKey,
